将courseware实体中将size的类型修改成了long类型 完成了课件上传下载(下载响应还是utf-8编码)模块

+++ easyclass-domain/src/main/java/edu/vero/easyclass/domain/Courseware.java
@@ -21,7 +21,7 @@ public class Courseware
 
     private String filePath;
 
-    private int size;
+    private long size;
 
     private TeacherArrangement arrangement;
 
@@ -68,13 +68,11 @@ public class Courseware
         this.filePath = filePath;
     }
 
-    public int getSize()
-    {
+    public long getSize() {
         return size;
     }
 
-    public void setSize(int size)
-    {
+    public void setSize(long size) {
         this.size = size;
     }
